+++ src/extensions/scenario/scenarioext.cpp
@@ -945,8 +945,10 @@ static Cell Clip_Move(Cell cell, FacingType facing, int dist)
  * 
  *  @author: 06/09/1995 BRR - Red Alert source code.
  *           CCHyper - Adjustments for Tiberian Sun.
+ * 
+ *  #issue-338 - Adds "min_dist" argument.
  */
-static int Scan_Place_Object(ObjectClass *obj, Cell cell)
+static int Scan_Place_Object(ObjectClass *obj, Cell cell, int min_dist = 1, int max_dist = 31, bool no_scatter = false)
 {
     int dist;               // for object placement
     FacingType rot;         // for object placement
@@ -978,7 +980,7 @@ static int Scan_Place_Object(ObjectClass *obj, Cell cell)
      *  If that fails, go to the next distance.
      *  This ensures that the closest coordinates are filled first.
      */
-    for (dist = 1; dist < 32; dist++) {
+    for (dist = min_dist; dist <= max_dist; dist++) {
 
         /**
          *  Pick a random starting direction
@@ -1006,7 +1008,7 @@ static int Scan_Place_Object(ObjectClass *obj, Cell cell)
                  *  If this is our second try at this distance, add a random scatter
                  *  to the desired cell, so our units aren't all aligned along spokes.
                  */
-                if (tryval > 0) {
+                if (!no_scatter && tryval > 0) {
                     newcell = Clip_Scatter(newcell, 1);
                 }
 
@@ -1040,6 +1042,92 @@ static int Scan_Place_Object(ObjectClass *obj, Cell cell)
                 if (rot > FACING_NW) {
                     rot = FACING_N;
                 }
+            }
+        }
+    }
+
+    return false;
+}
+
+
+/**
+ *  Checks if the cell adjacent from the input cell is occupied.
+ * 
+ *  @author: CCHyper
+ */
+static bool Is_Adjacent_Cell_Empty(Cell cell, FacingType facing, int dist)
+{
+    Cell newcell;
+    TechnoClass *techno;
+
+    /**
+     *  Pick a coordinate along this directional axis
+     */
+    newcell = Clip_Move(cell, facing, dist);
+
+    /**
+     *  Is there already an object on this cell?
+     */
+    techno = Map[newcell].Cell_Techno();
+    if (!techno) {
+        return true;
+    }
+    
+    /**
+     *  Is there any free infantry spots?
+     */
+    if (techno->What_Am_I() == RTTI_INFANTRY
+        && Map[newcell].Is_Any_Spot_Free()) {
+
+        return true;
+    }
+
+    return false;
+}
+
+
+static bool Are_Starting_Cells_Full(Cell cell, int dist)
+{
+    static bool empty_flag[FACING_COUNT];
+    std::memset(empty_flag, false, FACING_COUNT);
+
+    for (FacingType facing = FACING_FIRST; facing < FACING_COUNT; ++facing) {
+        if (Is_Adjacent_Cell_Empty(cell, facing, dist)) {
+            return false;
+        }
+    }
+
+    return true;
+}
+
+
+/**
+ *  Places an object >at< the given cell.
+ * 
+ *  @author: CCHyper
+ * 
+ *  #issue-338 - Adds "min_dist" argument.
+ */
+static bool Place_Object(ObjectClass *obj, Cell cell, FacingType facing, int dist)
+{
+    Cell newcell;
+    TechnoClass *techno;
+
+    /**
+     *  Pick a coordinate along this directional axis
+     */
+    newcell = Clip_Move(cell, facing, dist);
+
+    /**
+     *  Try to unlimbo the object in the given cell.
+     */
+    if (Map.In_Radar(newcell)) {
+        techno = Map[newcell].Cell_Techno();
+        if (!techno) {
+            Coordinate coord = Cell_Coord(newcell, true);
+            coord.Z = Map.Get_Cell_Height(coord);
+            if (obj->Unlimbo(coord, DIR_N)) {
+                return true;
             }
         }
     }
@@ -1118,7 +1206,16 @@ static DynamicVectorClass<Cell> Build_Starting_Waypoint_List(bool official)
  */
 void ScenarioClassExtension::Create_Units(bool official)
 {
-    DynamicVectorClass<TechnoClass *> deployed_objects;
+    /**
+     *  #issue-338
+     * 
+     *  Change the starting unit formation to be like Red Alert 2.
+     * 
+     *  This sets the desired placement distance from the base center cell.
+     * 
+     *  @author: CCHyper
+     */
+    const unsigned int PLACEMENT_DISTANCE = 3;
 
     int tot_units = Session.Options.UnitCount;
     if (Session.Options.Bases) {
@@ -1136,16 +1233,14 @@ void ScenarioClassExtension::Create_Units(bool official)
     /**
      *  Generate lists of all the available starting units (regardless of owner).
      */
-    int tot_count = 0;
-    int tot_cost = 0;
-    int budget = 0;
+    int tot_inf_count = 0;
+    int tot_unit_count = 0;
 
     for (int i = 0; i < UnitTypes.Count(); ++i) {
         UnitTypeClass *unittype = UnitTypes[i];
         if (unittype && unittype->IsAllowedToStartInMultiplayer) {
             if (Rule->BaseUnit->Fetch_ID() != unittype->Fetch_ID()) {
-                tot_cost += unittype->Raw_Cost();
-                ++tot_count;
+                ++tot_unit_count;
             }
         }
     }
@@ -1153,21 +1248,12 @@ void ScenarioClassExtension::Create_Units(bool official)
     for (int i = 0; i < InfantryTypes.Count(); ++i) {
         InfantryTypeClass *infantrytype = InfantryTypes[i];
         if (infantrytype && infantrytype->IsAllowedToStartInMultiplayer) {
-            tot_cost += infantrytype->Raw_Cost();
-            ++tot_count;
+            ++tot_inf_count;
         }
     }
 
-    if (!tot_count) {
+    if (!(tot_inf_count + tot_unit_count)) {
         DEBUG_WARNING("No starting units available!");
-    }
-
-    /**
-     *  #BUGFIX:
-     *  Check to prevent division by zero crash.
-     */
-    if (tot_cost != 0 && tot_count != 0) {
-        budget = tot_units * (tot_cost / tot_count);
     }
 
     /**
@@ -1350,7 +1436,7 @@ void ScenarioClassExtension::Create_Units(bool official)
              * 
              *  @author: CCHyper
              */
-            if (SessionExtension->ExtOptions.IsPrePlacedConYards) {
+            if (SessionExtension && SessionExtension->ExtOptions.IsPrePlacedConYards) {
 
                 /**
                  *  Create a construction yard (decided from the base unit).
@@ -1378,7 +1464,7 @@ void ScenarioClassExtension::Create_Units(bool official)
                          */
                         if (Session.Type != GAME_NORMAL) {
 
-                            if (!building->House->Is_Human_Control()) {
+                            if (!building->House->Is_Player_Control()) {
 
                                 building->IsToRebuild = true;
                                 building->IsToRepair = true;
@@ -1429,7 +1515,7 @@ void ScenarioClassExtension::Create_Units(bool official)
                          *  @author: CCHyper
                          */
                         if (Session.Options.UnitCount == 1) {
-                            if (SessionExtension->ExtOptions.IsAutoDeployMCV) {
+                            if (SessionExtension && SessionExtension->ExtOptions.IsAutoDeployMCV) {
                                 if (hptr->Is_Human_Control()) {
                                     obj->Set_Mission(MISSION_UNLOAD);
                                 }
@@ -1446,86 +1532,206 @@ void ScenarioClassExtension::Create_Units(bool official)
         }
 
         /**
-         *  Clear the previous house's deployed list.
+         *  #BUGFIX:
+         *  Make sure there are units available to place before entering the loop.
          */
-        deployed_objects.Clear();
+        bool units_available = (tot_inf_count + tot_unit_count) > 0;
 
-        //DEBUG_INFO("  budget == %d\n", budget);
+        if (units_available) {
 
-        if (budget > 0) {
+            TechnoTypeClass *technotype = nullptr;
+
+            int inf_percent = 50;
+            int unit_percent = 50;
+
+            int inf_count = (Session.Options.UnitCount * inf_percent) / 100;
+            int unit_count = (Session.Options.UnitCount * unit_percent) / 100;
 
             /**
-             *  Calculate the cost cap for units.
+             *  Make sure we place 3 infantry per cell.
              */
-            int unit_cost_cap = 2 * budget / 3;
+            inf_count *= 3;
 
             /**
              *  Place starting units for this house.
              */
-            int i = 0;
-            while (i < budget) {
+            if (available_units.Count() > 0) {
+                for (int i = 0; i < unit_count; ++i) {
 
-                TechnoTypeClass *technotype = nullptr;
+                    /**
+                     *  #BUGFIX:
+                     *  If all cells are full, we can stop placing units. This
+                     *  stops any run away cases with Scan_Place_Object.
+                     */
+                    if (Are_Starting_Cells_Full(centroid, PLACEMENT_DISTANCE)) {
+                        break;
+                    }
 
-                if (i < unit_cost_cap && available_units.Count() > 0) {
                     technotype = available_units[Random_Pick(0, available_units.Count()-1)];
-                } else if (available_infantry.Count() > 0) {
-                    technotype = available_infantry[Random_Pick(0, available_infantry.Count()-1)];
-                }
-
-                if (!technotype) {
-                    DEBUG_WARNING("  Invalid techno pointer!\n");
-                    continue;
-                }
-
-                /**
-                 *  Create an instance of the unit.
-                 */
-                obj = reinterpret_cast<TechnoClass *>(technotype->Create_One_Of(hptr));
-                if (obj) {
-
-                    if (Scan_Place_Object(obj, centroid)) {
-
-                        DEBUG_INFO("  House %s deployed object %s at %d,%d\n",
-                            hptr->Class->Name(), obj->Name(), obj->Get_Cell().X, obj->Get_Cell().Y);
-
-                        i += technotype->Raw_Cost();
-
-                        if (Scen->SpecialFlags.IsInitialVeteran) {
-                            obj->Veterancy.Set_Elite(true);
-                        }
-
-                        if (hptr->Is_Human_Control()) {
-                            obj->Set_Mission(MISSION_GUARD);
-                        } else {
-                            obj->Set_Mission(MISSION_GUARD_AREA);
-                        }
-
-                        /**
-                         *  Add to the list of objects successfully deployed.
-                         */
-                        deployed_objects.Add(obj);
-
-                    } else if (obj) {
-                        delete obj;
+                    if (!technotype) {
+                        DEBUG_WARNING("  Invalid unit pointer!\n");
+                        continue;
                     }
 
-                }
+                    /**
+                     *  Create an instance of the unit.
+                     */
+                    obj = reinterpret_cast<TechnoClass *>(technotype->Create_One_Of(hptr));
+                    if (obj) {
 
-                /**
-                 *  Scatter all the human placed objects to create
-                 *  some space around the base unit.
-                 */
-                if (hptr->Is_Human_Control()) {
-                    for (int i = 0; i < deployed_objects.Count(); ++i) {
-                        TechnoClass *techno = deployed_objects[i];
-                        if (techno) {
-                            techno->Scatter();
+                        if (Scan_Place_Object(obj, centroid, PLACEMENT_DISTANCE, PLACEMENT_DISTANCE, true)) {
+
+                            DEBUG_INFO("  House %s deployed object %s at %d,%d\n",
+                                hptr->Class->Name(), obj->Name(), obj->Get_Cell().X, obj->Get_Cell().Y);
+
+                            if (Scen->SpecialFlags.IsInitialVeteran) {
+                                obj->Veterancy.Set_Elite(true);
+                            }
+
+                            if (hptr->Is_Human_Control()) {
+                                obj->Set_Mission(MISSION_GUARD);
+                            } else {
+                                obj->Set_Mission(MISSION_GUARD_AREA);
+                            }
+
+                        } else if (obj) {
+                            delete obj;
                         }
+
                     }
+
                 }
 
             }
+
+            /**
+             *  Place starting infantry for this house.
+             */
+            if (available_infantry.Count() > 0) {
+                for (int i = 0; i < inf_count; ++i) {
+
+                    /**
+                     *  #BUGFIX:
+                     *  If all cells are full, we can stop placing units. This
+                     *  stops any run away cases with Scan_Place_Object.
+                     */
+                    if (Are_Starting_Cells_Full(centroid, PLACEMENT_DISTANCE)) {
+                        break;
+                    }
+
+                    technotype = available_infantry[Random_Pick(0, available_infantry.Count()-1)];
+                    if (!technotype) {
+                        DEBUG_WARNING("  Invalid infantry pointer!\n");
+                        continue;
+                    }
+
+                    /**
+                     *  Create an instance of the unit.
+                     */
+                    obj = reinterpret_cast<TechnoClass *>(technotype->Create_One_Of(hptr));
+                    if (obj) {
+
+                        if (Scan_Place_Object(obj, centroid, PLACEMENT_DISTANCE, PLACEMENT_DISTANCE, true)) {
+
+                            DEBUG_INFO("  House %s deployed object %s at %d,%d\n",
+                                hptr->Class->Name(), obj->Name(), obj->Get_Cell().X, obj->Get_Cell().Y);
+
+                            if (Scen->SpecialFlags.IsInitialVeteran) {
+                                obj->Veterancy.Set_Elite(true);
+                            }
+
+                            if (hptr->Is_Human_Control()) {
+                                obj->Set_Mission(MISSION_GUARD);
+                            } else {
+                                obj->Set_Mission(MISSION_GUARD_AREA);
+                            }
+
+                        } else if (obj) {
+                            delete obj;
+                        }
+
+                    }
+
+                }
+
+            }
+
+            /**
+             *  #issue-338
+             * 
+             *  Change the starting unit formation to be like Red Alert 2.
+             *  As a result, this is no longer required as the units are
+             *  now placed neatly around the base unit.
+             * 
+             *  @author: CCHyper
+             */
+#if 0
+            /**
+             *  Scatter all the human placed objects to create
+             *  some space around the base unit.
+             */
+            if (hptr->Is_Human_Control()) {
+                for (int i = 0; i < deployed_objects.Count(); ++i) {
+                    TechnoClass *techno = deployed_objects[i];
+                    if (techno) {
+                        techno->Scatter();
+                    }
+                }
+            }
+#endif
+
+#if 0
+            /**
+             *  #BUGFIX:
+             * 
+             *  Due to the costings of the starting units in Tiberian Sun, sometimes
+             *  there was a deficiency in the equal placement of units in the radius
+             *  around the starting unit. This code makes sure there are no blank
+             *  spaces around the base unit and that all players get 9 units.
+             */
+            if (Session.Options.UnitCount) {
+                for (FacingType facing = FACING_FIRST; facing < FACING_COUNT; ++facing) {
+                    if (Is_Adjacent_Cell_Empty(centroid, facing, PLACEMENT_DISTANCE)) {
+
+                        TechnoTypeClass *technotype = nullptr;
+
+                        /**
+                         *  Very rarely should another unit be placed, the algorithm
+                         *  above places a fair amount already...
+                         */
+                        if (Percent_Chance(25)) {
+                            technotype = available_units[Random_Pick(0, available_units.Count()-1)];
+                        } else if (available_infantry.Count() > 0) {
+                            technotype = available_infantry[Random_Pick(0, available_infantry.Count()-1)];
+                        }
+
+                        /**
+                         *  Create an instance of the unit.
+                         */
+                        obj = reinterpret_cast<TechnoClass *>(technotype->Create_One_Of(hptr));
+                        if (obj) {
+                            if (Place_Object(obj, centroid, facing, PLACEMENT_DISTANCE)) {
+                                DEBUG_WARNING("  House %s deployed deficiency object %s at %d,%d\n",
+                                    hptr->Class->Name(), obj->Name(), obj->Get_Cell().X, obj->Get_Cell().Y);
+
+                                if (Scen->SpecialFlags.InitialVeteran) {
+                                    obj->Veterancy.Set_Elite(true);
+                                }
+
+                                if (hptr->Is_Human_Control()) {
+                                    obj->Set_Mission(MISSION_GUARD);
+                                } else {
+                                    obj->Set_Mission(MISSION_GUARD_AREA);
+                                }
+
+                            } else if (obj) {
+                                delete obj;
+                            }
+                        }
+                    }
+                }
+            }
+#endif
 
         }
     }
